refactor: enhance SettingsRow structure and color scheme management
- Modularized SettingsRow by extracting components into SettingsRowContent for better organization.
- Updated color scheme application to prevent theme flash by using a dedicated script.
- Improved color scheme selection logic in SettingsScreen to utilize appColorSchemeAt for better index handling.
- Cleaned up unused styles and components to streamline the codebase.
