+++ app/(tabs)/(more)/index.tsx
@@ -239,7 +239,7 @@ const MoreScreen = () => {
           title: 'المساعدة',
           description: 'الأسئلة الشائعة والمساعدة والدعم',
           icon: <HelpSVG width={22} height={22} style={iconStyle} />,
-          onPress: handleOpenHelp,
+          onPress: () => void handleOpenHelp(),
         },
       ],
     },
@@ -275,7 +275,7 @@ const MoreScreen = () => {
           title: 'شارك التطبيق',
           description: 'شارك التطبيق مع أصدقائك',
           icon: <ShareSVG width={22} height={22} style={iconStyle} />,
-          onPress: handleShare,
+          onPress: () => void handleShare(),
         },
       ],
     },
